Ex4 Updated Document and Factory

+++ translator/src/document/coreDocument.java
@@ -1,6 +1,6 @@
 package document;
 
 abstract public class coreDocument extends manageDocument {
-    protected int size = 0;
+    protected byte size = 0;
 }
